Update patch set 4
Patch Set 4:

> Installing the aodh and ceilometer tests make this works, if this
 > is going to be the inteded usage this would have to be *really*
 > well documented.

To expand on this: if we take a base CentOS machine and install openstack-tempest, it works fine. However, installing python-aodh but *not* python-aodh-tests creates the issue in the fpaste. Just installing the -tests subpackage fixes it (so it is the expected behaviour).

As Daniel said, we need to clearly document the different options:

- openstack-tempest-all: tempest + all test packages
- openstack-tempest: just tempest. You need to manually install any -tests package for the installed services

Patch-set: 4
Label: Code-Review=0