Update patch set 1
Patch Set 1:

I have no idea about the current plans of moving u-c; maybe a reminder to the PTL would be useful.

Yes, they are assumed to be bug free, this was an especially complicated issue, difficult to discover (we found it in really special conditions, with a specific plugin that implemented some features, when the endpoints are in a special way, and so on), and required a bit of trickery to be fixed. It is more an exception and there is no way around: apply the patches or bump tempest to the last master.

Patch-set: 1
Reviewer: Gerrit User 141 <141@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0