Disable warning-is-error for sphinx
Without https://review.openstack.org/#/c/523235/ tarballs fail to build
because sphinx toc links to plugin-registry page that is generated from
git (we disable generation in .spec because it requires network
connectivity / git repo access).

This is a workaround until warning-is-error is disabled in upstream, or
some other solution for GENERATE_TEMPEST_PLUGIN_LIST=False users is
implemented.

Change-Id: Ia987bdc41ed0fdcf64171c98a77d901d3a29e440