Update patch set 1
Patch Set 1:

> My question is: if we remove the testrepository dependency,
 > wouldn't the generated package for 17.2 fail when used? I didn't
 > inspect the full chain of dependencies though.

 > My question is: if we remove the testrepository dependency,
 > wouldn't the generated package for 17.2 fail when used? I didn't
 > inspect the full chain of dependencies though.

For RDO Trunk packages, we'll be fine, because we're using master branch for Queens [1]. This won't be an issue for stable packages once https://review.openstack.org/546906 lands and we build 18.0.0, I think.

For Pike, we're using 17.2.0, so that's where we should not cherry-pick this patch.

[1] - https://github.com/redhat-openstack/rdoinfo/blob/master/rdo.yml#L4708-L4719

Patch-set: 1
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0