Replace Recommends with Requires for the -all subpackage
https://review.rdoproject.org/r/#/c/25320/ moved the spec file
to Recommends which leads to making all the dependencies to weak
dependencies and during container build with in dnf config, weak
deps are set to false so all these dependencies are not installed
in containers leading to missing dependencies.

Switching to Requires fixes the issue.

The change in rpm-master made by:
https://review.rdoproject.org/r/#/c/27962

Change-Id: I5b40c8259e6c37b10b85575247f6fe0bf23fdbc6