Added workaround for handling mock py2 & py3 issue
In order to package tempest for Train release which supports
also py2, we need to add a workaround which replaces all
'from unittest import mock' by 'import mock' as in py2 mock is
not a part of unittest.
Also a new dependency configparser must be installed as the
configparser from six.moves lib got deprecated and a patch
which removed using _future_ module must be revert for py2
compatibility.

Change-Id: If4a38b9e73ace59e7b3c818142c90512d10ec837