Update patch set 19
Patch Set 19:

oh, the problem with patches instead of sed commands is that the build fails if the patch it not applicable. The patches are needed in tempest 25 however in the context of this review the tempest is still at 24 tag. The bump is made by this review https://review.rdoproject.org/r/#/c/29611/ which won't pass without the patches.

Patch-set: 19
Reviewer: Gerrit User 266 <266@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0