Revert tempest/manager.py removal
This reverts removal of an old tempest/manager.py in order to
provide compatibility with neutron-tempest-plugin which has
switched to the new manager location at very latest and the
change hasn't been included in a tag nor a release. We revert
the removal in tempest so that we can package tempest 26.1.0
and ship it in Victoria. Tempest 26.1.0 contains apart from
the fixes also a significant argument deprecation.

Change-Id: Iec8ef16a117998d1be4791b4c55ccb1569981a1d