Backport relevant wait_until_sshable_pingable patches
https://review.opendev.org/q/topic:wait_until_sshable_pingable
The above patches are needed to close following bugs:
- https://bugs.launchpad.net/tripleo/+bug/1960310
- https://bugzilla.redhat.com/show_bug.cgi?id=2012096

Below is the list of reviews:
- https://review.opendev.org/c/openstack/tempest/+/843155
- https://review.opendev.org/c/openstack/tempest/+/817771
- https://review.opendev.org/c/openstack/tempest/+/817772

These patches are proactively backported to Downstream.
In order to keep the spec file in sync, we are backporting
those patches here also.

It will also close following bugs:
- https://bugs.launchpad.net/tripleo/+bug/1978553
- https://bugs.launchpad.net/tripleo/+bug/1978556
which got detected during the above series backport.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I1a0a0d5e1a67389707f30bbfa3f22eaf2c7e654e